requirements.txt
torch
matplotlib
loguru
tiktoken
datasets
python-dotenv
huggingface-hub